_ai/ml-service/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# ML Libraries
scikit-learn==1.3.2
tensorflow==2.15.0
xgboost==2.0.2

# Data Processing
pandas==2.1.3
numpy==1.26.2
psycopg2-binary==2.9.9

# Utilities
joblib==1.3.2
python-dateutil==2.8.2
pytz==2023.3

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
